🏗️ Refactor to replace worker by HostedService
